backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pandas==2.3.0
numpy==2.2.6
scikit-learn==1.7.0
python-dotenv==1.1.0
pydantic==2.11.5